+++ src/cadqueryhelper/__init__.py
@@ -3,3 +3,4 @@ from .irregularGrid import irregular_grid
 from .Base import Base
 from .Hinge import Hinge
 from .HelperPlate import HelperPlate
+from .randomizedRotationGrid import randomized_rotation_grid