+++ changes.md
@@ -19,6 +19,7 @@
 * Formalized the pipe shape and added parameters.
 * Added vase shape and example / documentation
 * Added HelperPlate
+* Updated README.md
 
 
 ## 0.2.1